chore: subscribe pre subscribe
